requirements.txt
# Checkers game engine
imparaai-checkers
pytest==5.4.1
fastapi==0.54.1
uvicorn==0.11.3
aiohttp
